feat: unify hanlder of plain-text
