+++ packages/core/src/docs/data-model/document-data-model.ts
@@ -15,8 +15,13 @@
  */
 
 import { BehaviorSubject } from 'rxjs';
-import type { Nullable } from '../../shared';
+import { UnitModel, UniverInstanceType } from '../../common/unit';
 import { Tools } from '../../shared/tools';
+import { getEmptySnapshot } from './empty-snapshot';
+import { JSONX } from './json-x/json-x';
+import { PRESET_LIST_TYPE } from './preset-list-type';
+import { getBodySlice, SliceBodyType } from './text-x/utils';
+import type { Nullable } from '../../shared';
 import type {
     IDocumentBody,
     IDocumentData,
@@ -26,12 +31,7 @@ import type {
     IListData,
 } from '../../types/interfaces/i-document-data';
 import type { IPaddingData } from '../../types/interfaces/i-style-data';
-import { UnitModel, UniverInstanceType } from '../../common/unit';
-import { getBodySlice, SliceBodyType } from './text-x/utils';
-import { getEmptySnapshot } from './empty-snapshot';
 import type { JSONXActions } from './json-x/json-x';
-import { JSONX } from './json-x/json-x';
-import { PRESET_LIST_TYPE } from './preset-list-type';
 
 export const DEFAULT_DOC = {
     id: 'default_doc',
