+++ packages/sheets-data-validation/src/utils/get-cell-data-origin.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { DataStreamTreeTokenType, DEFAULT_EMPTY_DOCUMENT_VALUE, type ICellData, type Nullable } from '@univerjs/core';
+import { BuildTextUtils, type ICellData, type Nullable } from '@univerjs/core';
 
 export function getCellValueOrigin(cell: Nullable<ICellData>) {
     if (cell === null) {
@@ -29,10 +29,8 @@ export function getCellValueOrigin(cell: Nullable<ICellData>) {
         }
 
         const data = body.dataStream;
-        const lastString = data.substring(data.length - 2, data.length);
-        const newDataStream = lastString === DEFAULT_EMPTY_DOCUMENT_VALUE ? data.substring(0, data.length - 2) : data;
-
-        return newDataStream.replaceAll(DataStreamTreeTokenType.CUSTOM_RANGE_START, '').replaceAll(DataStreamTreeTokenType.CUSTOM_RANGE_END, '');
+        const newDataStream = BuildTextUtils.transform.getPlainText(data);
+        return newDataStream;
     }
 
     return cell?.v;
